O:  rename model path
