+++ leaderboard/pendubot/robustness_v2/con_ar_eapo.py
@@ -13,7 +13,7 @@ leaderboard_config = {
 }
 
 model_path = Path(
-    "../../../data/policies/design_C.1/model_1.1/pendubot/AR-EAPO/model.zip"
+    "../../../data/policies/design_C.1/model_1.1/pendubot/AR_EAPO/model.zip"
 )
 controller = AR_EAPOController(
     model_path=model_path,
